Add custom charts example and clean up
