+++ vizro-core/docs/pages/user_guides/custom_components.md
@@ -144,7 +144,7 @@ vm.Parameter.add_type("selector", TooltipNonCrossRangeSlider)
 
     === "yaml"
         ```yaml
-        # TBD
+        # Custom components are currently only possible via python configuration
         ```
     === "Result"
         [![CustomComponent1]][CustomComponent1]
@@ -264,7 +264,7 @@ vm.Page.add_type("components", Jumbotron)
         6.  The new component can now be inserted into a regular dashboard.
     === "yaml"
         ```yaml
-        # TBD
+        # Custom components are currently only possible via python configuration
         ```
     === "Result"
         [![CustomComponent2]][CustomComponent2]
